+++ src/db.rs
@@ -5,8 +5,8 @@ use crate::execution::{build_write, try_collect};
 use crate::expression::function::scala::ScalarFunctionImpl;
 use crate::expression::function::table::TableFunctionImpl;
 use crate::expression::function::FunctionSummary;
-use crate::function::characterlength::CharacterLength;
-use crate::function::charlength::CharLength;
+use crate::function::char_length::CharLength;
+use crate::function::char_length::CharacterLength;
 use crate::function::current_date::CurrentDate;
 use crate::function::lower::Lower;
 use crate::function::numbers::Numbers;
@@ -53,7 +53,7 @@ impl DataBaseBuilder {
             table_functions: Default::default(),
         };
         builder = builder.register_scala_function(CharLength::new());
-        builder = builder.register_scala_function(CharacterLength::new());
+        builder = builder.register_scala_function(CharacterLength::new2());
         builder = builder.register_scala_function(CurrentDate::new());
         builder = builder.register_scala_function(Lower::new());
         builder = builder.register_scala_function(Upper::new());
