+++ src/function/mod.rs
@@ -1,5 +1,4 @@
-pub(crate) mod characterlength;
-pub(crate) mod charlength;
+pub(crate) mod char_length;
 pub(crate) mod current_date;
 pub(crate) mod lower;
 pub(crate) mod numbers;
